Added rough note placement Incorrectly scaled, moving all quads along the piano roll

+++ Piano/PianoEngine/src/core/global_stats.h
@@ -8,8 +8,8 @@
 
 static struct
 {
-  double realTime; // Real time in seconds since the main loop began
-  double deltaTime; // Real time in milliseconds of the previous frame
+  f32 realTime; // Real time in seconds since the main loop began
+  f32 deltaTime; // Real time in milliseconds of the previous frame
 } globalStats;
 
 #endif // !define PIANO_CORE_GLOBAL_STATS_H_
